Split out BhotmEntry into multiple components, standardized CSS class format
